Update README.md
Description
This program is a simple contract written in Solidity, a programming language used for developing smart contracts on the Ethereum blockchain. The contract has a three function one for mapping, second for mining and third for burning the token.

The mapping function maps the address with balances.

The mint function is used to mint the nft and add the minted amount to total_supply and balances of the address provided.

The burn function is used to burn the nft that we have minted and subtract the burn amount from total_supply and balances of the address provided. Here, we have provided a if statement to check balance is greater than the value to be burnt, if balance is less than value than the further code will not execute.

Getting Started
Executing Program
To run this program, you can use Remix, an online Solidity IDE. To get started, go to the Remix website at https://remix.ethereum.org/.

Once you are on the Remix website, create a new file by clicking on the "+" icon in the left-hand sidebar. Save the file with a .sol extension (e.g., error.sol). Copy and paste the following code written by me into the file.

To compile the code, click on the "Solidity Compiler" tab in the left-hand sidebar. Make sure the "Compiler" option is set heigher to "0.8.1" (or another compatible version), and then click on the "Compile error.sol" button.

Once the code is compiled, you can deploy the contract by clicking on the "Deploy & Run Transactions" tab in the left-hand sidebar. Select the "MyToken" contract from the dropdown menu, and then click on the "Deploy" button.

Once the contract is deployed, you can interact with it by passing the address. call any of the three function and set the value, and you can burn, mint or check the balance of your NFT.

Author
Aman Patel

License
This project is licensed under the MIT License - see the LICENSE file for details